Redesigned service regulator & drop WalletServiceLegacy
