+++ config/config.php
@@ -34,7 +34,6 @@ use Bavix\Wallet\Services\ExchangeService;
 use Bavix\Wallet\Services\PrepareService;
 use Bavix\Wallet\Services\PurchaseService;
 use Bavix\Wallet\Services\RegulatorService;
-use Bavix\Wallet\Services\StateService;
 use Bavix\Wallet\Services\TaxService;
 
 return [
@@ -81,7 +80,6 @@ return [
         'basket' => BasketService::class,
         'bookkeeper' => BookkeeperService::class,
         'regulator' => RegulatorService::class,
-        'state' => StateService::class,
         'cast' => CastService::class,
         'consistency' => ConsistencyService::class,
         'discount' => DiscountService::class,
