+++ src/Traits/CanConfirm.php
@@ -20,7 +20,6 @@ use Bavix\Wallet\Services\AtomicServiceInterface;
 use Bavix\Wallet\Services\CastServiceInterface;
 use Bavix\Wallet\Services\ConsistencyServiceInterface;
 use Bavix\Wallet\Services\RegulatorServiceInterface;
-use Bavix\Wallet\Services\StateServiceInterface;
 use Illuminate\Database\RecordsNotFoundException;
 
 trait CanConfirm
@@ -79,7 +78,6 @@ trait CanConfirm
 
             $wallet = app(CastServiceInterface::class)->getWallet($this);
             app(RegulatorServiceInterface::class)->decrease($wallet, $transaction->amount);
-            app(StateServiceInterface::class)->persist($wallet);
 
             return $transaction->update(['confirmed' => false]);
         });
@@ -122,7 +120,6 @@ trait CanConfirm
             }
 
             app(RegulatorServiceInterface::class)->increase($wallet, $transaction->amount);
-            app(StateServiceInterface::class)->persist($wallet);
 
             return $transaction->update(['confirmed' => true]);
         });
